pyserver/pyproject.toml
[project]
name = "trilogy-studio-servers"
version = "0.1.0"
description = "A simple language service for Trilogy"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "efromvt" }]
license = { text = "MIT" }
dependencies = [
    "pytrilogy",
    "anyio>=4.5",
    "click>=8.1.0",
    "httpx>=0.27",
    "mcp[cli]",
    "pydantic>=2.0",
    "pydantic-settings>=2.5.2",
    "sse-starlette>=1.6.1",
    "uvicorn>=0.23.1; sys_platform != 'emscripten'",
]
